code/backend/requirements.txt
flask
flask-cors
pillow
opencv-python
numpy
requests
google-cloud-vision
